Mavenified klangfarbe
