+++ src/main/java/com/github/klangfarbe/statechart/FinalState.java
@@ -1,0 +1,34 @@
+/*
+ * UML statechart framework (http://github.com/klangfarbe/UML-Statechart-Framework-for-Java)
+ *
+ * Copyright (C) 2006-2013 Christian Mocek (christian.mocek@gmail.com)
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ * 
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
+ */
+package com.github.klangfarbe.statechart;
+
+/**
+ * The final state of a statechart or substate
+ */
+public class FinalState extends State {
+    /**
+     * Creates a new Finalstate.
+     * 
+     * @throws StatechartException
+     */
+    public FinalState(String name, Context parent) throws StatechartException {
+        super(name, parent, null, null, null);
+    };
+}
