--- unittests/statechart/unittests/TestAction.java
@@ -1,42 +1,0 @@
-/*
- * UML statechart framework (http://github.com/klangfarbe/UML-Statechart-Framework-for-Java)
- *
- * Copyright (C) 2006-2013 Christian Mocek (christian.mocek@gmail.com)
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- * 
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
- */
-package statechart.unittests;
-
-import statechart.Action;
-import statechart.Metadata;
-import statechart.Parameter;
-
-public class TestAction implements Action {
-  public TestAction(String name, String action) {
-    this.name = name;
-    this.action = action;
-  }
-
-  public void execute(Metadata data, Parameter param) {
-    TestParameter parameter = (TestParameter) param;
-    parameter.path += 
-        (parameter.path.length() != 0 ? " " : "") 
-        + action + ":" + name;
-  };
-
-  private String name;
-
-  private String action;
-}
